Ver 1.0.1 Bugfix: fixed == and != operators (now compares the double given in input with the double cast of num/den). New: added support for full matrices. New: added Back_Substitution() method for finding solution in full matrices. New: updated main() with new features.

+++ Numbers.cpp
@@ -188,7 +188,7 @@ bool Fraction::operator<=(const Fraction &right) {
 
 bool Fraction::operator==(const double& right) {
     double fraction;
-    fraction = num/den;
+    fraction = (float)num/den;
     if(fraction == right)
         return true;
     else
@@ -197,7 +197,7 @@ bool Fraction::operator==(const double& right) {
 
 bool Fraction::operator!=(const double& right) {
     double fraction;
-    fraction = num/den;
+    fraction = (float)num/den;
     if(fraction != right)
         return true;
     else
